Add belt-and-braces check for MOS.bin file size
I noticed that with the test for the ZDI interface not working,
the code carries on, giving the impression of doing something,
causing great confusion. When this happened, I noticed that it
thought the filesize was 16777215 (0xFFFFFF). Check for this as
an extra validation step. With the fix for the ZDI interface
it shouldn't be necessary, but I think it is worth keeping.